refactor: separate runtime services from coordinators
