+++ src/yoyopod/coordinators/__init__.py
@@ -5,8 +5,14 @@ Coordinator modules for YoyoPod orchestration.
 from yoyopod.coordinators.call import CallCoordinator
 from yoyopod.coordinators.power import PowerCoordinator
 from yoyopod.coordinators.playback import PlaybackCoordinator
-from yoyopod.coordinators.runtime import AppRuntimeState, CoordinatorRuntime
+from yoyopod.coordinators.registry import AppRuntimeState, CoordinatorRuntime
 from yoyopod.coordinators.screen import ScreenCoordinator
+from yoyopod.coordinators.voice import (
+    VoiceCommandExecutor,
+    VoiceCommandOutcome,
+    VoiceRuntimeCoordinator,
+    VoiceSettingsResolver,
+)
 
 __all__ = [
     "AppRuntimeState",
@@ -15,4 +21,8 @@ __all__ = [
     "PlaybackCoordinator",
     "CoordinatorRuntime",
     "ScreenCoordinator",
+    "VoiceCommandExecutor",
+    "VoiceCommandOutcome",
+    "VoiceRuntimeCoordinator",
+    "VoiceSettingsResolver",
 ]
