+++ src/yoyopod/coordinators/call.py
@@ -11,7 +11,7 @@ from typing import Optional
 
 from loguru import logger
 
-from yoyopod.coordinators.runtime import AppRuntimeState, CoordinatorRuntime
+from yoyopod.coordinators.registry import AppRuntimeState, CoordinatorRuntime
 from yoyopod.coordinators.screen import ScreenCoordinator
 from yoyopod.core import EventBus
 from yoyopod.core import (
