+++ src/yoyopod/coordinators/playback.py
@@ -8,7 +8,7 @@ from loguru import logger
 
 from yoyopod.audio.music import LocalMusicService
 from yoyopod.audio.music.models import Track
-from yoyopod.coordinators.runtime import AppRuntimeState, CoordinatorRuntime
+from yoyopod.coordinators.registry import AppRuntimeState, CoordinatorRuntime
 from yoyopod.coordinators.screen import ScreenCoordinator
 from yoyopod.core import EventBus
 from yoyopod.core import (
