+++ tests/test_call_coordinator.py
@@ -8,7 +8,7 @@ from pathlib import Path
 from yoyopod.core import AppContext
 from yoyopod.communication import CallHistoryStore, CallState, RegistrationState
 from yoyopod.coordinators.call import CallCoordinator
-from yoyopod.coordinators.runtime import CoordinatorRuntime
+from yoyopod.coordinators.registry import CoordinatorRuntime
 from yoyopod.core import CallFSM, CallInterruptionPolicy, MusicFSM
