+++ tests/test_fsm_runtime.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 import yoyopod.core.fsm.music as fsm_module
-from yoyopod.coordinators.runtime import AppRuntimeState, CoordinatorRuntime
+from yoyopod.coordinators.registry import AppRuntimeState, CoordinatorRuntime
 from yoyopod.core import (
     CallFSM,
     CallInterruptionPolicy,
