+++ tests/test_runtime_voice.py
@@ -1,4 +1,4 @@
-"""Direct tests for the shared runtime-owned voice orchestration seam."""
+"""Direct tests for the shared voice coordination seam."""
 
 from __future__ import annotations
 
@@ -8,7 +8,7 @@ from tempfile import NamedTemporaryFile
 from types import SimpleNamespace
 
 from yoyopod.core import AppContext
-from yoyopod.runtime.voice import (
+from yoyopod.coordinators.voice import (
     VoiceCommandExecutor,
     VoiceCommandOutcome,
     VoiceRuntimeCoordinator,
@@ -328,7 +328,7 @@ def test_voice_runtime_coordinator_releases_cached_service_when_settings_change(
         def release_resources(self) -> None:
             self.released = True
 
-    monkeypatch.setattr("yoyopod.runtime.voice.coordinator.VoiceManager", _TrackingVoiceManager)
+    monkeypatch.setattr("yoyopod.coordinators.voice.coordinator.VoiceManager", _TrackingVoiceManager)
     coordinator = VoiceRuntimeCoordinator(
         context=None,
         settings_resolver=VoiceSettingsResolver(
